Fix: copy text 함수 위치 변경
클립보드에 텍스트 저장하는 함수 위치 변경
